Adjust the javascript files..
