+++ index.html
@@ -265,7 +265,7 @@
       </form>
     </section>
   </main>
-  <script type="module" src="js/script.js"></script>
+  <script src="js/script.js"></script>
 
 </body>
